refactor: getProjectPublicInfo -> getProjectInfoSummary로 변경
