+++ src/app/post/[slug]/page.tsx
@@ -2,10 +2,12 @@
 
 import PostDetail from '@/features/post/public/contents/postDetail/PostDetail';
 
-export default function PostPage({
+const PostPage = ({
   params: { slug: postId },
 }: {
   params: { slug: string };
-}) {
+}) => {
   return <PostDetail postId={postId} />;
-}
+};
+
+export default PostPage;
