+++ src/features/project/private/components/myProjects/ProjectCard.tsx
@@ -2,7 +2,7 @@ import { format } from 'date-fns';
 import Link from 'next/link';
 import TechStackImage from '@/components/techStack/public/TechStackImage';
 import { FaPlusCircle } from '@react-icons/all-files/fa/FaPlusCircle';
-import { ProjectInfoSummary } from '@/service/project/public/getProjectPublicInfo';
+import { ProjectInfoSummary } from '@/service/project/public/getProjectInfoSummary';
 
 interface ProjectCardProps {
   projectPost: ProjectInfoSummary;
